Windows/MSVC portability: fix psapi link, long-truncation, path detection, C1061, and long double precision divergence
Following up on the tusdview/textools MSVC fixes, audited src/next,
src/tydra/next, and legacy src/ core for the same class of Windows
portability issues:

- src/tydra/next/CMakeLists.txt: link psapi PUBLIC on tydra_next for
  Windows (GetProcessMemoryInfo in mem-budget.hh), mirroring the existing
  tools/tusdrender fix. Missing link would break any target linking
  tydra_next directly (e.g. test_tydra_next) on MSVC.
- src/tydra/next/render-converter.cc: drop unnecessary long() casts around
  three std::reverse calls -- long is 32-bit on Windows/LLP64 (64-bit on
  Linux/macOS/LP64), a silent-truncation risk for very large meshes that
  vector::iterator's difference_type doesn't need.
- src/tydra/next/texture-cache.cc: reuse AssetResolver::IsAbsolutePath
  instead of a POSIX-only `path[0] != '/'` check, so Windows drive-absolute
  (C:\...) and UNC (\server\share) asset paths aren't incorrectly treated
  as base_dir-relative.
- src/next/pipeline/flatten.cc: route FileSizeBytes through SafeStreamSize
  like every other file-size probe in src/next, instead of raw tellg().
- src/composition.cc: remove a dead __linux__-only <unistd.h> include (no
  symbols from it are used).
- src/stage-converter.cc: refactor CrateWriter::ConvertValue's ~150-branch
  type_name else-if chain into standalone-if TRY_CONVERT_SET/SET_NOROLE/
  UNREG macros, mirroring the TRY_PACK_DICT pattern already used in
  crate-writer-values.cc for the same reason -- the else-if chain had
  already hit MSVC C1061 ("blocks nested too deeply") once and was only
  defensively (fragile, ad hoc) split into 4 segments. Verified
  behavior-preserving: the multiset of type_name string literals compared
  is byte-identical before and after (161 occurrences incl. 6 pre-existing
  dead/duplicate branches, 155 unique, preserved as-is).
- src/next/eval/value-clip.cc, src/value-clip-utils.hh: both
  ExpandTemplate{,ClipMetadata} implementations (next-core and legacy-core)
  used `long double` for overflow/tolerance guard headroom that only
  exists on GCC/Clang/x86 (80-bit extended precision) -- MSVC's long
  double is bit-identical to double, so the same USD file could
  round/reject differently per compiler. Rewrote both in plain double with
  an explicit overflow-safe-add helper, correct by construction on every
  platform.
- src/tydra/next/mem-budget.hh: same long double -> double cleanup for a
  GiB-to-bytes overflow guard (double's precision is already far more than
  enough margin here).

Verified: tinyusdz_static, tydra_next, tinyusdz_next, and tusdview all
rebuild clean (no C1061, no new warnings). feat-value-clip,
usdc-roundtrip-test, and usdc-parser-unit-test pass. A standalone harness
reproducing unit-value-clips.cc's 4 cases (blocked from running via ctest
by an unrelated pre-existing near-macro collision bug in a sibling test
file) passes against the rewritten value-clip-utils.hh with byte-identical
output.

+++ src/tydra/next/mem-budget.hh
@@ -60,11 +60,17 @@ class MemBudget {
   // cap_override_gib <= 0 -> the shared 32 GiB-target host policy.
   void Init(double cap_override_gib) {
     if (cap_override_gib > 0.0) {
-      const long double requested =
-          static_cast<long double>(cap_override_gib) *
-          static_cast<long double>(size_t(1) << 30);
-      const long double max_size = static_cast<long double>(
-          (std::numeric_limits<size_t>::max)());
+      // Plain `double` here, not `long double`: MSVC's `long double` is
+      // bit-identical to `double` (no 80-bit extended precision like
+      // GCC/Clang/x86 get), and double's ~15-17 significant decimal digits
+      // are already far more than enough to place `requested` correctly on
+      // either side of `max_size` (~1.8e19, i.e. SIZE_MAX on 64-bit) -- no
+      // realistic --max-mem/--max-vram GiB value comes close to that
+      // boundary, so no compiler-specific extra precision is needed here.
+      const double requested =
+          cap_override_gib * static_cast<double>(size_t(1) << 30);
+      const double max_size =
+          static_cast<double>((std::numeric_limits<size_t>::max)());
       cap_ = requested >= max_size
                  ? (std::numeric_limits<size_t>::max)()
                  : static_cast<size_t>(requested);
